% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_prod.R
\name{cross_prod}
\alias{cross_prod}
\title{Prediction using a tidyfit data frame}
\arguments{
\item{fit}{a tibble of fitted models generated with \code{tidyfit}.}

\item{data}{a data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr). Must have the same structure as the input data to \code{tidyfit}.}
}
\value{
A \code{tibble}.
}
\description{
This function performs a prediction using a tidyfit data frame
}
\details{
Prediction is performed by multiplying the data frame with the coefficient vectors trained using \code{tidyfit}.
}
\examples{
data <- tidyfit::Factor_Industry_Returns
fit <- regress(data, Return ~ ., lin_reg = m("lm"), .mask = "Date")
cross_prod(fit, data)

}
\seealso{
\code{\link{regress}} and \code{\link{classify}} methods
}
\author{
Johann Pfitzinger
}
