% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.dbscan.R
\name{.fit.dbscan}
\alias{.fit.dbscan}
\title{DBSCAN for \code{tidyfit}}
\usage{
\method{.fit}{dbscan}(self, data = NULL)
}
\arguments{
\item{self}{a 'tidyFit' R6 class.}

\item{data}{a data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}
}
\value{
A fitted 'tidyFit' class model.
}
\description{
Calculates clusters using density-based spatial clustering of applications with noise (DBSCAN) on a 'tidyFit' \code{R6} class. The function can be used with \code{\link{cluster}}.
}
\details{
\strong{Hyperparameters:}

\emph{None. Cross validation not applicable.}

\strong{Important method arguments (passed to \code{\link{m}})}
\itemize{
\item k (number of clusters)
\item eps
\item minPts
}

The function provides a wrapper for \code{dbscan::dbscan}. See \code{?dbscan} for more details.

\strong{Implementation}

The formula can be specified with a LHS variable, as \code{~ x + z}.
}
\examples{
# Load data
data <- tidyfit::Factor_Industry_Returns[1:100,]

# Stand-alone function
fit <- m("dbscan",  ~ SMB + HML, data)
fit

# Within 'cluster' function
fit <- cluster(data, ~ ., m("dbscan", eps = 1), .mask = c("Date", "Industry"))
fitted(fit)

# Cluster across columns
fit <- cluster(data, ~ ., m("dbscan", eps = 1, transpose = TRUE), .mask = c("Date", "Industry"))
fitted(fit)

}
\seealso{
\code{\link{.fit.diana}}, \code{\link{.fit.dtw}}, \code{\link{.fit.agnes}} and \code{\link{m}} methods
}
\author{
Johann Pfitzinger
}
