% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.cor.R
\name{.model.cor}
\alias{.model.cor}
\title{Pearson correlation coefficients for \code{tidyfit}}
\usage{
\method{.model}{cor}(x = NULL, y = NULL, control = NULL, ...)
}
\arguments{
\item{x}{Input matrix or data.frame, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.}

\item{y}{Response variable.}

\item{control}{Additional arguments passed to \code{cor}.}

\item{...}{Not used.}
}
\value{
A 'tibble'.
}
\description{
Calculates Pearson correlation coefficients and returns the results as a tibble. The function can be used with \code{\link{regress}}.
}
\details{
\strong{Hyperparameters:}

\emph{None. Cross validation not applicable.}

The function provides a wrapper for \code{stats::cor}.

Note that this method is not compatible with \code{\link{cross_prod}} and the results will be nonsensical.
}
\examples{
x <- matrix(rnorm(100 * 20), 100, 20)
y <- rnorm(100)
fit <- m("cor", x, y)
fit

}
\seealso{
\code{\link{m}} method
}
\author{
Johann Pfitzinger
}
