% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.lm.R
\name{.model.lm}
\alias{.model.lm}
\title{Linear regression for \code{tidyfit}}
\usage{
\method{.model}{lm}(formula = NULL, data = NULL, control = NULL, ...)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted.}

\item{data}{a data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{control}{Additional arguments passed to \code{lm}.}

\item{...}{Not used.}
}
\value{
A 'tibble'.
}
\description{
Fits a linear regression and returns the results as a tibble. The function can be used with \code{\link{regress}}.
}
\details{
\strong{Hyperparameters:}

\emph{None. Cross validation not applicable.}

The function provides a wrapper for \code{stats::lm}.

An argument \code{vcov.} can be passed in control or to \code{...} in \code{\link{m}} to estimate the model with robust standard errors. \code{vcov.} can be one of "BS", "HAC", "HC" and "OPG" and is passed to the \code{sandwich} package.
}
\examples{
# Load data
data <- tidyfit::Factor_Industry_Returns

# Stand-alone function
fit <- m("lm", Return ~ ., data)
fit

# Within 'regress' function
fit <- regress(data, Return ~ ., m("lm"), .mask = c("Date", "Industry"))
coef(fit)

# With robust standard errors
fit <- m("lm", x, y, vcov. = "HAC")
fit

}
\seealso{
\code{\link{.model.robust}}, \code{\link{.model.glm}} and \code{\link{m}} methods
}
\author{
Johann Pfitzinger
}
