% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.mslm.R
\name{.model.mslm}
\alias{.model.mslm}
\title{Markov-Switching Regression for \code{tidyfit}}
\usage{
\method{.model}{mslm}(formula = NULL, data = NULL, control = NULL, ...)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted.}

\item{data}{a data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{control}{Additional arguments passed to \code{MSwM::msmFit}.}

\item{...}{Not used.}
}
\value{
A 'tibble'.
}
\description{
Fits a Markov-Switching regression and returns the results as a tibble. The function can be used with \code{\link{regress}}.
}
\details{
\strong{Hyperparameters:}

\emph{None. Cross validation not applicable.}

The function provides a wrapper for \code{MSwM::msmFit}. Note that only the regression method with 'lm' is implemented at this stage.

An argument \code{index_col} can be passed, which allows a custom index to be added to \code{coef(m("tvp"))} (e.g. a date index). This only works when the formula is specified as \code{y ~ .} (with a dot on the RHS), since otherwise the index_col is currently removed by \code{\link{regress}}.
}
\examples{
# Load data
data <- tidyfit::Factor_Industry_Returns
data <- dplyr::filter(data, Industry == "HiTec")
data <- dplyr::select(data, -Industry)

ctr <- list(maxiter = 10, parallelization = FALSE)

# Stand-alone function
fit <- m("mslm", Return ~ ., data, index_col = "Date", k = 2, control = ctr)
fit

# Within 'regress' function
fit <- regress(data, Return ~ .,
               m("mslm", index_col = "Date", k = 2, control = ctr))
tidyr::unnest(coef(fit), model_info)

}
\seealso{
\code{\link{.model.tvp}} and \code{\link{m}} methods
}
\author{
Johann Pfitzinger
}
