% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.pcr.R
\name{.model.pcr}
\alias{.model.pcr}
\title{Principal Components Regression for \code{tidyfit}}
\usage{
\method{.model}{pcr}(formula = NULL, data = NULL, control = NULL, ...)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted.}

\item{data}{a data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{control}{Additional arguments passed to \code{pls::pcr}.}

\item{...}{Not used.}
}
\value{
A 'tibble'.
}
\description{
Fits a principal components regression and returns the results as a tibble. The function can be used with \code{\link{regress}}.
}
\details{
\strong{Hyperparameters:}
\itemize{
\item \code{ncomp} \emph{(number of components)}
}

The principal components regression is fitted using \code{pls} package. Covariates are standardized, with coefficients back-transformed to the original scale. An intercept is always included.

If no hyperparameter grid is passed (\code{is.null(control$ncomp)}), the default is \code{seq(1, nvars)}, where nvars is the number of features.

Note that at present \code{pls} does not offer weighted implementations or non-gaussian response. The method can therefore only be used with \code{\link{regress}}
}
\examples{
# Load data
data <- tidyfit::Factor_Industry_Returns

# Stand-alone function
fit <- m("pcr", Return ~ ., data, ncomp = 3)
fit

# Within 'regress' function
fit <- regress(data, Return ~ ., m("pcr"), .mask = c("Date", "Industry"), .cv = "vfold")
coef(fit)

}
\references{
Kristian Hovde Liland, Bjørn-Helge Mevik, Ron Wehrens (2022).
pls: Partial Least Squares and Principal Component Regression.
R package version 2.8-1. URL https://CRAN.R-project.org/package=pls.
}
\seealso{
\code{\link{.model.plsr}} and \code{\link{m}} methods
}
\author{
Johann Pfitzinger
}
