% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.spikeslab.R
\name{.model.spikeslab}
\alias{.model.spikeslab}
\title{Bayesian Spike and Slab regression for \code{tidyfit}}
\usage{
\method{.model}{spikeslab}(self, data = NULL)
}
\arguments{
\item{self}{a tidyFit R6 class.}

\item{data}{a data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}
}
\value{
A fitted tidyFit class model.
}
\description{
Fits a Bayesian Spike and Slab regression on a 'tidyFit' \code{R6} class. The function can be used with \code{\link{regress}}.
}
\details{
\strong{Hyperparameters:}

\emph{None. Cross validation not applicable.}

\strong{Important method arguments (passed to \code{\link{m}})}
\itemize{
\item expected.r2
\item prior.df
\item expected.model.size
\item niter
}

The function provides a wrapper for \code{BoomSpikeSlab::lm.spike}. See \code{?lm.spike} for more details.

\strong{Implementation}

Prior arguments are passed to BoomSpikeSlab::SpikeSlabPrior (the function automatically identifies which arguments are for the prior, and which for BoomSpikeSlab::lm.spike).
}
\examples{
# Load data
data <- tidyfit::Factor_Industry_Returns

# Stand-alone function
fit <- m("spikeslab", Return ~ ., data, niter = 100)
fit

# Within 'regress' function
fit <- regress(data, Return ~ ., m("spikeslab", niter = 100),
               .mask = c("Date", "Industry"))
coef(fit)

}
\references{
Scott SL (2022). \emph{BoomSpikeSlab: MCMC for Spike and Slab Regression}. R package version 1.2.5, \url{https://CRAN.R-project.org/package=BoomSpikeSlab}.
}
\seealso{
\code{\link{.model.lasso}}, \code{\link{.model.blasso}} and \code{\link{m}} methods
}
\author{
Johann Pfitzinger
}
