% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.tvp.R
\name{.model.tvp}
\alias{.model.tvp}
\title{Bayesian Time-Varying Regression for \code{tidyfit}}
\usage{
\method{.model}{tvp}(formula = NULL, data = NULL, control = NULL, ...)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted.}

\item{data}{a data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{control}{Additional arguments passed to \code{lm}.}

\item{...}{Not used.}
}
\value{
A 'tibble'.
}
\description{
Fits a Bayesian time-varying regression and returns the results as a tibble. The function can be used with \code{\link{regress}}.
}
\details{
\strong{Hyperparameters:}

\emph{None. Cross validation not applicable.}

The function provides a wrapper for \code{shrinkTVP::shrinkTVP}.

An argument \code{index_col} can be passed, which allows a custom index to be added to \code{coef(m("tvp"))} (e.g. a date index). This only works when the formula is specified as \code{y ~ .} (with a dot on the RHS), since otherwise the index_col is currently removed by \code{\link{regress}}.
}
\examples{
# Load data
data <- tidyfit::Factor_Industry_Returns
data <- dplyr::filter(data, Industry == "HiTec")
data <- dplyr::select(data, -Industry)

# Stand-alone function
fit <- m("tvp", Return ~ ., data, index_col = "Date", niter = 100)
fit

# Within 'regress' function
fit <- regress(data, Return ~ ., m("tvp", niter = 100, index_col = "Date"))
tidyr::unnest(coef(fit), model_info)

}
\seealso{
\code{\link{.model.robust}}, \code{\link{.model.glm}} and \code{\link{m}} methods
}
\author{
Johann Pfitzinger
}
