% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain}
\alias{explain}
\title{An interface for variable importance measures for a fitted tidyfit.models frames}
\usage{
explain(object, use_package = NULL, use_method = NULL, ...)
}
\arguments{
\item{object}{\code{model.frame} created using \code{\link{regress}}, \code{\link{classify}} or \code{\link{m}}}

\item{use_package}{the package to use to calculate variable importance. See 'Details' for possible options.}

\item{use_method}{the method from 'use_package' that should be used to calculate variable importance.}

\item{...}{additional arguments passed to the importance method}
}
\value{
A 'tibble'.
}
\description{
A generic method for calculating XAI and variable importance methods for tidyfit.models frames.
}
\details{
\strong{WARNING}
This function is currently in an experimental stage.

The function uses the 'model_object' column in a \code{tidyfit.model} frame to return variable importance measures for each model.

\strong{Possible packages and methods include:}
\subsection{\code{sensitivity} package:}{

The package provides methods to assess variable importance in linear regressions ('lm') and classifications ('glm').

\emph{Usage:} \code{use_package="sensitivity"}
\emph{Methods:}
\itemize{
\item "lmg" (Shapley regression),
\item "pmvd" (Proportional marginal variance decomposition),
\item "src" (standardized regression coefficients),
\item "pcc" (partial correlation coefficients),
\item "johnson" (Johnson indices)
}

See \code{?sensitivity::lmg} for more information and additional arguments.
}

\subsection{\code{iml} package:}{

Integration with iml is currently in progress. The methods can be used for 'nnet', 'rf', 'lasso', 'enet', 'ridge', 'adalasso', 'glm' and 'lm'.

\emph{Usage:} \code{use_package="iml"}
\emph{Methods:}
\itemize{
\item "Shapley" (SHAP values)
\item "LocalModel" (LIME)
\item "FeatureImp" (Permutation-based feature importance)
}

The argument 'which_rows' (vector of integer indexes) can be used to explain specific rows in the data set for Shapley and LocalModel methods.
}

\subsection{\code{randomForest} package:}{

This uses the native importance method of the randomForest package and can be used with 'rf' and 'quantile_rf' regression and classification.

\emph{Usage:} \code{use_package="randomForest"}
\emph{Methods:}
\itemize{
\item "mean_decrease_accuracy"
}
}
}
\examples{
data <- dplyr::group_by(tidyfit::Factor_Industry_Returns, Industry)
fit <- regress(data, Return ~ ., m("lm"), .mask = "Date")
explain(fit, use_package = "sensitivity", use_method = "src")

data <- dplyr::filter(tidyfit::Factor_Industry_Returns, Industry == Industry[1])
fit <- regress(data, Return ~ ., m("lm"), .mask = c("Date", "Industry"))
explain(fit, use_package = "iml", use_method = "Shapley", which_rows = c(1))

}
\references{
Molnar C, Bischl B, Casalicchio G (2018). “iml: An R package for Interpretable Machine Learning.” \emph{JOSS},
\emph{3}(26), 786. \url{doi:10.21105/joss.00786}.

Iooss B, Veiga SD, Janon A, Pujol G, Broto wcfB, Boumhaout K, Clouvel L, Delage T, Amri RE, Fruth J, Gilquin
L, Guillaume J, Herin M, Idrissi MI, Le Gratiet L, Lemaitre P, Marrel A, Meynaoui A, Nelson BL, Monari F,
Oomen R, Rakovec O, Ramos B, Rochet P, Roustant O, Sarazin G, Song E, Staum J, Sueur R, Touati T, Verges V,
Weber F (2024). \emph{sensitivity: Global Sensitivity Analysis of Model Outputs and Importance Measures}. R
package version 1.30.0, \url{https://CRAN.R-project.org/package=sensitivity}.

A. Liaw and M. Wiener (2002). Classification and Regression by randomForest. R News 2(3), 18--22.
}
\author{
Johann Pfitzinger
}
