% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.tidyfit.models.R
\name{explain.tidyfit.models}
\alias{explain.tidyfit.models}
\title{Explain details of a fitted tidyfit.models frame}
\usage{
\method{explain}{tidyfit.models}(object, method = NULL, ..., .keep_grid_id = FALSE)
}
\arguments{
\item{object}{\code{model.frame} created using \code{\link{regress}}, \code{\link{classify}} or \code{\link{m}}}

\item{method}{the variable importance method used to create explanations. See 'Details' for possible options.}

\item{...}{additional arguments passed to the importance method}
}
\value{
A 'tibble'.
}
\description{
A generic method for calculating XAI and variable importance methods for tidyfit.models frames.
}
\details{
The function uses the 'model_object' column in a \code{tidyfit.model} frame to return variable importance measures for each model.

\strong{Possible methods include:}
\subsection{Linear regression ('lm'):}{
\itemize{
\item 'shapley_reg' for Shapley regression (default). The method uses the 'lmg' algorithm implemented in the \code{relaimpo}-package. Other algorithms can be applied by explicitly passing 'type' to \code{...}, which is passed to \code{relaimpo::calc.relimp}.
\item 'rel_weights' for relative weights. The method uses the 'genizi' algorithm implemented in the \code{relaimpo}-package.
}
}
}
\examples{
data <- dplyr::group_by(tidyfit::Factor_Industry_Returns, Industry)
fit <- regress(data, Return ~ ., m("lm"), .mask = "Date")
explain(fit, method = "rel_weights")

}
\author{
Johann Pfitzinger
}
