% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Factor_Industry_Returns}
\alias{Factor_Industry_Returns}
\title{Industry-Factor Returns Data Set}
\description{
The data set includes monthly returns between 1963 and 2022 for 10 industries, as well as factor values for 5 Fama-French factors.
}
\references{
\url{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\keyword{data}
