% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_model}
\alias{get_model}
\title{Get a fitted model from a tidyfit.models frame}
\usage{
get_model(df, row = 1)
}
\arguments{
\item{df}{a tidyfit.models frame created using m(), regress(), classify() and similar methods}

\item{row}{an integer index of the row number in 'df' for which the model should be returned}
}
\value{
An object of the class associated with the underyling fitting algorithm
}
\description{
Returns a single fitted model object produced by the underlying fitting algorithm from a tidyfit.models frame based on a given row number.
}
\details{
This method is a utility to return the object fitted by the underlying algorithm. For instance, when \code{m("lm")} is used to create the tidyfit.models frame, the returned object is of class "lm".
}
\examples{
# Load data
data <- tidyfit::Factor_Industry_Returns

# Stand-alone function
fit <- m("lm", Return ~ ., data)
summary(get_model(fit))

}
\seealso{
\code{\link{get_tidyFit}} method
}
\author{
Johann Pfitzinger
}
