% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_tidyFit}
\alias{get_tidyFit}
\title{Get a tidyFit model from a tidyfit.models frame}
\usage{
get_tidyFit(df, row = 1)
}
\arguments{
\item{df}{a tidyfit.models frame created using m(), regress(), classify() and similar methods}

\item{row}{an integer index of the row number in 'df' for which the model should be returned}
}
\value{
An object of the class associated with the underyling fitting algorithm
}
\description{
Returns a single tidyFit object from a tidyfit.models frame based on a given row number.
}
\details{
This method is a utility to return the tidyFit object from a row index of the tidyfit.models frame. The tidyFit object contains the fitted model and several additional objects necessary to reproduce the analysis or refit the model on new data.
}
\examples{
# Load data
data <- tidyfit::Factor_Industry_Returns

# Stand-alone function
fit <- m("lm", Return ~ ., data)
model <- get_tidyFit(fit)

# Refit
model$fit(data)

}
\seealso{
\code{\link{get_model}} method
}
\author{
Johann Pfitzinger
}
