% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.boost.R
\name{m.boost}
\alias{m.boost}
\title{Gradient boosting regression for tidyfit}
\usage{
m.boost(x, y, mstop = NULL, ...)
}
\arguments{
\item{x}{input matrix or data.frame, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.}

\item{y}{response variable.}

\item{mstop}{an integer giving the number of initial boosting iterations. If mstop = 0, the offset model is returned.}

\item{...}{Additional arguments passed to \code{mboost::glmboost}.}
}
\value{
A 'tibble'.
}
\description{
Fits a gradient boosting regression and returns the results as a tibble. The function can be used with \code{tidyfit}.
}
\details{
The gradient boosting regression is performed using 'mboost::glmboost'.

The function can be used for classification or regression, covariates are standardized and an intercept is always included.
}
\examples{
x = matrix(rnorm(100 * 20), 100, 20)
y = rbinom(100, 1, 0.5)
fit = m.boost(x, y, mstop = 100, family = "binomial")

}
\references{
T. Hothorn, P. Buehlmann, T. Kneib, M. Schmid, and B. Hofner (2022).
mboost: Model-Based Boosting, R package version 2.9-7, https://CRAN.R-project.org/package=mboost.
}
\seealso{
\code{tidypredict} method
}
\author{
Johann Pfitzinger
}
