% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.cor.R
\name{m.cor}
\alias{m.cor}
\title{Pearson correlation coefficients for tidyfit}
\usage{
m.cor(x, y, ...)
}
\arguments{
\item{x}{Input matrix or data.frame, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.}

\item{y}{Response variable.}

\item{...}{Additional arguments passed to \code{cor}.}
}
\value{
A 'tibble'.
}
\description{
Calculates Pearson correlation coefficients and returns the results as a tibble. The function can be used with \code{tidyfit}.
}
\details{

}
\examples{
x = matrix(rnorm(100 * 20), 100, 20)
y = rnorm(100)
fit = m.cor(x, y)
fit

}
\references{

}
\seealso{
\code{tidypredict} method
}
\author{
Johann Pfitzinger
}
