% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.enet.R
\name{m.enet}
\alias{m.enet}
\title{ElasticNet regression or classification for tidyfit}
\usage{
m.enet(x, y, lambda = NULL, alpha = NULL, ...)
}
\arguments{
\item{x}{Input matrix or data.frame, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.}

\item{y}{Response variable.}

\item{lambda}{shrinkage parameter or vector of shrinkage parameters.}

\item{alpha}{The elasticnet mixing parameter. \code{alpha=1} is the lasso penalty, and \code{alpha=0} the ridge penalty.}

\item{...}{Additional arguments passed to \code{glmnet}.}
}
\value{
A 'tibble'.
}
\description{
Fits an elasticnet regression or classification and returns the results as a tibble. The function can be used with \code{tidyfit}.
}
\details{

}
\examples{
x = matrix(rnorm(100 * 20), 100, 20)
y = rnorm(100)
fit = m.enet(x, y, kappa = 0.5)
fit

}
\references{

}
\seealso{
\code{tidypredict} method
}
\author{
Johann Pfitzinger
}
