% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.glm.R
\name{m.glm}
\alias{m.glm}
\title{Logistic regression for tidyfit}
\usage{
m.glm(x, y, ...)
}
\arguments{
\item{x}{Input matrix or data.frame, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.}

\item{y}{Response variable.}

\item{...}{Additional arguments passed to \code{lm}.}
}
\value{
A 'tibble'.
}
\description{
Fits a logistic regression and returns the results as a tibble. The function can be used with \code{tidyfit}.
}
\details{

}
\examples{
x = matrix(rnorm(100 * 20), 100, 20)
y = rbinom(100, 1, 0.5)
fit = m.glm(x, y)
fit

}
\references{

}
\seealso{
\code{m.lm} method
}
\author{
Johann Pfitzinger
}
