% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.hfr.R
\name{m.hfr}
\alias{m.hfr}
\title{Hierarchical feature regression for tidyfit}
\usage{
m.hfr(x, y, kappa = NULL, ...)
}
\arguments{
\item{x}{Input matrix or data.frame, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.}

\item{y}{Response variable.}

\item{kappa}{shrinkage penalty or sequence of shrinkage penalties.}

\item{...}{Additional arguments passed to \code{hfr::cv.hfr}.}
}
\value{
A 'tibble'.
}
\description{
Fits a hierarchical feature regression and returns the results as a tibble. The function can be used with \code{regress}.
}
\details{

}
\examples{
x = matrix(rnorm(100 * 20), 100, 20)
y = rnorm(100)
fit = m.hfr(x, y, kappa = 0.5)
fit

}
\references{
Pfitzinger J (2022).
_hfr: Estimate Hierarchical Feature Regression Models_.
R package version 0.5.0, <https://CRAN.R-project.org/package=hfr>.
}
\seealso{
\code{tidypredict} method
}
\author{
Johann Pfitzinger
}
