% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.lasso.R
\name{m.lasso}
\alias{m.lasso}
\title{Lasso regression and classification for \code{tidyfit}}
\usage{
m.lasso(x, y, lambda = NULL, ...)
}
\arguments{
\item{x}{Input matrix or data.frame, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.}

\item{y}{Response variable.}

\item{lambda}{lambda value or lambda sequence for shrinkage penalty.}

\item{...}{Additional arguments passed to \code{lm}.}
}
\value{
A 'tibble'.
}
\description{
Fits a linear regression and returns the results as a tibble. The function can be used with \code{tidyfit}.
}
\details{

}
\examples{
x = matrix(rnorm(100 * 20), 100, 20)
y = rnorm(100)
fit = m.lasso(x, y, kappa = 0.5)
fit

}
\references{

}
\seealso{
\code{m.enet}, \code{m.ridge} and \code{m.adalasso} methods
}
\author{
Johann Pfitzinger
}
