% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.plsr.R
\name{m.plsr}
\alias{m.plsr}
\title{Partial Least Squares Regression for tidyfit}
\usage{
m.plsr(x, y, ncomp = NULL, ...)
}
\arguments{
\item{x}{Input matrix or data.frame, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.}

\item{y}{Response variable.}

\item{ncomp}{number of components or sequence of components.}

\item{...}{Additional arguments passed to \code{lm}.}
}
\value{
A 'tibble'.
}
\description{
Fits a partial least squares regression and returns the results as a tibble. The function can be used with \code{tidyfit}.
}
\details{
The partial least squares regression is fitted using \code{pls} package. Covariates are standardized, with coefficients back-transformed to the original scale.
}
\examples{
x = matrix(rnorm(100 * 20), 100, 20)
y = rnorm(100)
fit = m.plsr(x, y, ncomp = 4)
fit

}
\references{

}
\seealso{
\code{m.pcr} method
}
\author{
Johann Pfitzinger
}
