% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_boost.R
\name{mod_boost}
\alias{mod_boost}
\title{Gradient boosting regression for \code{tidyfit}}
\usage{
mod_boost(x = NULL, y = NULL, mstop = NULL, ...)
}
\arguments{
\item{x}{input matrix or data.frame, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.}

\item{y}{response variable.}

\item{mstop}{an integer giving the number of initial boosting iterations. If mstop = 0, the offset model is returned.}

\item{...}{Additional arguments passed to \code{mboost::glmboost}.}
}
\value{
A 'tibble'.
}
\description{
Fits a gradient boosting regression or classification and returns the results as a tibble. The function can be used with \code{regress}.
}
\details{
The gradient boosting regression is performed using 'mboost::glmboost'.

The function can be used for classification or regression, covariates are standardized and an intercept is always included.

If no hyperparameter grid is passed (\code{mstop = NULL}), the default grid is used with \code{mstop = c(100, 500, 1000, 2000)}.

When called without \code{x} and \code{y} arguments, the function returns a partialised version of the function that can be called with data to fit the model.
}
\examples{
x <- matrix(rnorm(100 * 20), 100, 20)
y <- rbinom(100, 1, 0.5)
fit <- mod_boost(x, y, mstop = 100, family = "binomial")
fit

}
\references{
T. Hothorn, P. Buehlmann, T. Kneib, M. Schmid, and B. Hofner (2022).
mboost: Model-Based Boosting, R package version 2.9-7, https://CRAN.R-project.org/package=mboost.
}
\seealso{
\code{tidypredict} method
}
\author{
Johann Pfitzinger
}
