% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cor.R
\name{mod_cor}
\alias{mod_cor}
\title{Pearson correlation coefficients for \code{tidyfit}}
\usage{
mod_cor(x = NULL, y = NULL, ...)
}
\arguments{
\item{x}{Input matrix or data.frame, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.}

\item{y}{Response variable.}

\item{...}{Additional arguments passed to \code{cor}.}
}
\value{
A 'tibble'.
}
\description{
Calculates Pearson correlation coefficients and returns the results as a tibble. The function can be used with \code{regress}.
}
\details{
When called without \code{x} and \code{y} arguments, the function returns a partialised version of the function that can be called with data to fit the model.
}
\examples{
x <- matrix(rnorm(100 * 20), 100, 20)
y <- rnorm(100)
fit <- mod_cor(x, y)
fit

}
\references{

}
\seealso{
\code{tidypredict} method
}
\author{
Johann Pfitzinger
}
