% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_enet.R
\name{mod_enet}
\alias{mod_enet}
\title{ElasticNet regression or classification for \code{tidyfit}}
\usage{
mod_enet(x = NULL, y = NULL, lambda = NULL, alpha = NULL, ...)
}
\arguments{
\item{x}{Input matrix or data.frame, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.}

\item{y}{Response variable.}

\item{lambda}{lambda value or lambda sequence for shrinkage penalty. See Details.}

\item{alpha}{The elasticnet mixing parameter. \code{alpha=1} is the lasso penalty, and \code{alpha=0} the ridge penalty. See Details.}

\item{...}{Additional arguments passed to \code{glmnet::glmnet}.}
}
\value{
A 'tibble'.
}
\description{
Fits an ElasticNet regression or classification and returns the results as a tibble. The function can be used with \code{regress}.
}
\details{
The ElasticNet regression is estimated using \code{glmnet::glmnet}. For classification pass \code{family = "binomial"} to \code{...}.

If no hyperparameter grid is passed (\code{lambda = NULL} and \code{alpha = NULL}), \code{dials::grid_regular()} is used to determine a sensible default grid. The grid size is 100 for \code{lambda} and 5 for \code{alpha}.

When called without \code{x} and \code{y} arguments, the function returns a partialised version of the function that can be called with data to fit the model.
}
\examples{
x <- matrix(rnorm(100 * 20), 100, 20)
y <- rnorm(100)
fit <- mod_enet(x, y, kappa = 0.5)
fit

}
\references{

}
\seealso{
\code{mod_lasso} and \code{mod_ridge} methods
}
\author{
Johann Pfitzinger
}
