% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_glm.R
\name{mod_glm}
\alias{mod_glm}
\title{Logistic regression for \code{tidyfit}}
\usage{
mod_glm(x = NULL, y = NULL, ...)
}
\arguments{
\item{x}{Input matrix or data.frame, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.}

\item{y}{Response variable.}

\item{...}{Additional arguments passed to \code{glm}.}
}
\value{
A 'tibble'.
}
\description{
Fits a logistic regression and returns the results as a tibble. The function can be used with \code{regress}.
}
\details{

}
\examples{
x <- matrix(rnorm(100 * 20), 100, 20)
y <- rbinom(100, 1, 0.5)
fit <- mod_glm(x, y)
fit

}
\references{

}
\seealso{
\code{mod_lm} method
}
\author{
Johann Pfitzinger
}
