% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hfr.R
\name{mod_hfr}
\alias{mod_hfr}
\title{Hierarchical feature regression for \code{tidyfit}}
\usage{
mod_hfr(x = NULL, y = NULL, kappa = NULL, ...)
}
\arguments{
\item{x}{Input matrix or data.frame, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.}

\item{y}{Response variable.}

\item{kappa}{shrinkage penalty or sequence of shrinkage penalties.}

\item{...}{Additional arguments passed to \code{hfr::cv.hfr}.}
}
\value{
A 'tibble'.
}
\description{
Fits a hierarchical feature regression and returns the results as a tibble. The function can be used with \code{regress}.
}
\details{
When called without \code{x} and \code{y} arguments, the function returns a partialised version of the function that can be called with data to fit the model.
}
\examples{
x <- matrix(rnorm(100 * 20), 100, 20)
y <- rnorm(100)
fit <- mod_hfr(x, y, kappa = 0.5)
fit

}
\references{
Pfitzinger J (2022).
_hfr: Estimate Hierarchical Feature Regression Models_.
R package version 0.5.0, <https://CRAN.R-project.org/package=hfr>.
}
\seealso{
\code{tidypredict} method
}
\author{
Johann Pfitzinger
}
