% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lm.R
\name{mod_lm}
\alias{mod_lm}
\title{Linear regression for \code{tidyfit}}
\usage{
mod_lm(
  x = NULL,
  y = NULL,
  vcov_type = c("default", "BS", "HAC", "HC", "OPG"),
  ...
)
}
\arguments{
\item{x}{Input matrix or data.frame, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.}

\item{y}{Response variable.}

\item{vcov_type}{Robust standard errors. \code{sandwich} is used to calculate the vcov using \code{sandwich::vcovHAC}, \code{sandwich::vcovBS} etc.}

\item{...}{Additional arguments passed to \code{lm}.}
}
\value{
A 'tibble'.
}
\description{
Fits a linear regression and returns the results as a tibble. The function can be used with \code{regress}.
}
\details{
When called without \code{x} and \code{y} arguments, the function returns a partialised version of the function that can be called with data to fit the model.
}
\examples{
x <- matrix(rnorm(100 * 20), 100, 20)
y <- rnorm(100)
fit <- mod_lm(x, y)
fit

# Partial execution
mod <- mod_lm(vcov = "HAC")
fit <- mod(x, y)
fit

}
\references{

}
\seealso{
\code{mod_lasso} method
}
\author{
Johann Pfitzinger
}
