% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pcr.R
\name{mod_pcr}
\alias{mod_pcr}
\title{Principal Components Regression for \code{tidyfit}}
\usage{
mod_pcr(x = NULL, y = NULL, ncomp = NULL, ...)
}
\arguments{
\item{x}{Input matrix or data.frame, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.}

\item{y}{Response variable.}

\item{ncomp}{number of components or sequence of components.}

\item{...}{Additional arguments passed to \code{pls::pcr}.}
}
\value{
A 'tibble'.
}
\description{
Fits a principal components regression and returns the results as a tibble. The function can be used with \code{regress}.
}
\details{
The principal components regression is fitted using \code{pls} package. Covariates are standardized, with coefficients back-transformed to the original scale.

If no hyperparameter grid is passed (\code{ncomps = NULL}), the default is \code{seq(1, nvars)}, where nvars is the number of features.

When called without \code{x} and \code{y} arguments, the function returns a partialised version of the function that can be called with data to fit the model.
}
\examples{
x <- matrix(rnorm(100 * 20), 100, 20)
y <- rnorm(100)
fit <- mod_pcr(x, y, ncomp = 4)
fit

}
\references{

}
\seealso{
\code{mod_plsr} method
}
\author{
Johann Pfitzinger
}
