% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ridge.R
\name{mod_ridge}
\alias{mod_ridge}
\title{Ridge regression and classification for \code{tidyfit}}
\usage{
mod_ridge(x = NULL, y = NULL, lambda = NULL, ...)
}
\arguments{
\item{x}{Input matrix or data.frame, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.}

\item{y}{Response variable.}

\item{lambda}{lambda value or lambda sequence for shrinkage penalty. See Details.}

\item{...}{Additional arguments passed to \code{glmnet::glmnet}.}
}
\value{
A 'tibble'.
}
\description{
Fits a linear regression or classification with L2 penalty and returns the results as a tibble. The function can be used with \code{regress}.
}
\details{
The ridge regression is estimated using \code{glmnet::glmnet} with \code{alpha = 0}. For classification pass \code{family = "binomial"} to \code{...}.

If no hyperparameter grid is passed (\code{lambda = NULL}), \code{dials::grid_regular()} is used to determine a sensible default grid. The grid size is 100.

When called without \code{x} and \code{y} arguments, the function returns a partialised version of the function that can be called with data to fit the model.
}
\examples{
x <- matrix(rnorm(100 * 20), 100, 20)
y <- rnorm(100)
fit <- mod_ridge(x, y)
fit

# Partial execution
mod <- mod_ridge(lambda = 1)
fit <- mod(x, y)
fit

}
\references{

}
\seealso{
\code{mod_enet}, \code{mod_lasso} and \code{mod_´adalasso} methods
}
\author{
Johann Pfitzinger
}
