% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyfit.R
\name{tidyfit}
\alias{tidyfit}
\title{Linear model estimation on tidy data}
\usage{
tidyfit(
  .data,
  formula,
  ...,
  .cv = c("none", "loo", "vfold"),
  .cv_args = list(v = 10),
  .hyper_grid = NULL,
  .weights = NULL,
  .mask = NULL
)
}
\arguments{
\item{...}{Additional arguments passed to \code{hclust}.}

\item{x}{Input matrix or data.frame, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.}
}
\value{
A \code{tibble}.
}
\description{
This function is a wrapper to fit many different types of linear
regression of classification model on a (potentially grouped) `tibble`.
}
\details{

}
\examples{

}
\references{

}
\seealso{
\code{tidypredict} method
}
\author{
Johann Pfitzinger
}
