% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_imp.tidyfit.models.R
\name{var_imp.tidyfit.models}
\alias{var_imp.tidyfit.models}
\title{Estimate variable importance from a \code{tidyfit.models} frame}
\usage{
\method{var_imp}{tidyfit.models}(object, ..., .keep_grid_id = FALSE)
}
\arguments{
\item{object}{\code{model.frame} created using \code{\link{regress}}, \code{\link{classify}} or \code{\link{m}}}

\item{...}{currently not used}

\item{.keep_grid_id}{boolean. By default the grid ID column is dropped, if there is only one unique setting per model or group. \code{.keep_grid_id = TRUE} ensures that the column is never dropped.}
}
\value{
A 'tibble'.
}
\description{
The function estimates relative feature importance from all models in a \code{tidyfit.models} frame and outputs a tidy frame.
}
\details{
Relative feature importance is estimated via a model agnostic method from the \code{iml} package (evaluating model error increase for feature permutation). This wrapper uses
\code{compare="ratio"} of \code{loss="mae"} for regression models, and \code{compare="difference"} of \code{loss="ce"} for classification models.
}
\examples{
data <- tidyfit::Factor_Industry_Returns
fit <- regress(data, Return ~ ., m("lm"), .mask = c("Date", "Industry"))
var_imp(fit)

}
\seealso{
\code{\link{predict.tidyfit.models}}, \code{\link{fitted.tidyfit.models}}, \code{\link{coef.tidyfit.models}} and \code{\link{residuals.tidyfit.models}}
}
\author{
Phil Holzmeister
}
