% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.bayes.R
\name{.model.bayes}
\alias{.model.bayes}
\title{Bayesian generalized linear regression for \code{tidyfit}}
\usage{
\method{.model}{bayes}(formula = NULL, data = NULL, control = NULL, ...)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted.}

\item{data}{a data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{control}{Additional arguments passed to \code{bayesglm}.}

\item{...}{Not used.}
}
\value{
A 'tibble'.
}
\description{
Fits a Bayesian regression and returns the results as a tibble. The function can be used with \code{\link{regress}} and \code{\link{classify}}.
}
\details{
\strong{Hyperparameters:}

\emph{None. Cross validation not applicable.}

The function provides a wrapper for \code{stats::glm}.
}
\examples{
# Load data
data <- tidyfit::Factor_Industry_Returns

# Stand-alone function
fit <- m("bayes", Return ~ ., data)
fit

# Within 'regress' function
fit <- regress(data, Return ~ ., m("bayes"), .mask = c("Date", "Industry"))
coef(fit)


}
\seealso{
\code{\link{.model.glm}} and \code{\link{m}} methods
}
\author{
Johann Pfitzinger
}
