% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.glm.R
\name{.model.glm}
\alias{.model.glm}
\title{Generalized linear regression for \code{tidyfit}}
\usage{
\method{.model}{glm}(x = NULL, y = NULL, control = NULL, ...)
}
\arguments{
\item{x}{Input matrix or data.frame, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.}

\item{y}{Response variable.}

\item{control}{Additional arguments passed to \code{glm}.}

\item{...}{Not used.}
}
\value{
A 'tibble'.
}
\description{
Fits a logistic regression and returns the results as a tibble. The function can be used with \code{\link{regress}} and \code{\link{classify}}.
}
\details{
\strong{Hyperparameters:}

\emph{None. Cross validation not applicable.}

The function provides a wrapper for \code{stats::glm}.
}
\examples{
x <- matrix(rnorm(100 * 20), 100, 20)
y <- rbinom(100, 1, 0.5)
fit <- m("glm", x, y, family = binomial(link = probit))
fit


}
\seealso{
\code{\link{.model.lm}} and \code{\link{m}} methods
}
\author{
Johann Pfitzinger
}
