% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.hfr.R
\name{.model.hfr}
\alias{.model.hfr}
\title{Hierarchical feature regression for \code{tidyfit}}
\usage{
\method{.model}{hfr}(x = NULL, y = NULL, control = NULL, ...)
}
\arguments{
\item{x}{Input matrix or data.frame, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.}

\item{y}{Response variable.}

\item{control}{Additional arguments passed to \code{hfr::cv.hfr}.}

\item{...}{Not used.}
}
\value{
A 'tibble'.
}
\description{
Fits a hierarchical feature regression and returns the results as a tibble. The function can be used with \code{\link{regress}}.
}
\details{
\strong{Hyperparameters:}
\itemize{
\item kappa (\emph{proportional size of regression graph})
}

The hierarchical feature regression is estimated using the \code{hfr::cv.hfr} function. An intercept is always included and features are standardized with coefficients transformed to the original scale.

If no hyperparameter grid is provided (\code{is.null(control$kappa)}), the default is \code{seq(0, 1, by = 0.1)}.
}
\examples{
x <- matrix(rnorm(100 * 20), 100, 20)
y <- rnorm(100)
fit <- m("hfr", x, y, kappa = 0.5)
fit

}
\references{
Pfitzinger J (2022).
\emph{hfr: Estimate Hierarchical Feature Regression Models}.
R package version 0.5.0, \url{https://CRAN.R-project.org/package=hfr}.
}
\seealso{
\code{\link{m}} method
}
\author{
Johann Pfitzinger
}
