% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.lm.R
\name{.model.lm}
\alias{.model.lm}
\title{Linear regression for \code{tidyfit}}
\usage{
\method{.model}{lm}(x = NULL, y = NULL, control = NULL, ...)
}
\arguments{
\item{x}{Input matrix or data.frame, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.}

\item{y}{Response variable.}

\item{control}{Additional arguments passed to \code{lm}.}

\item{...}{Not used.}
}
\value{
A 'tibble'.
}
\description{
Fits a linear regression and returns the results as a tibble. The function can be used with \code{\link{regress}}.
}
\details{
\strong{Hyperparameters:}

\emph{None. Cross validation not applicable.}

The function provides a wrapper for \code{stats::lm}.

An argument \code{vcov.} can be passed in control or to \code{...} in \code{\link{m}} to estimate the model with robust standard errors. \code{vcov.} can be one of "BS", "HAC", "HC" and "OPG" and is passed to the \code{sandwich} package.
}
\examples{
x <- matrix(rnorm(100 * 20), 100, 20)
y <- rnorm(100)
fit <- m("lm", x, y)
fit

# With robust standard errors
fit <- m("lm", x, y, vcov. = "HAC")
fit

}
\seealso{
\code{\link{.model.robust}}, \code{\link{.model.glm}} and \code{\link{m}} methods
}
\author{
Johann Pfitzinger
}
