% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.mslm.R
\name{.model.mslm}
\alias{.model.mslm}
\title{Markov-Switching Regression for \code{tidyfit}}
\usage{
\method{.model}{mslm}(self, data = NULL)
}
\arguments{
\item{self}{a tidyFit R6 class.}

\item{data}{a data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}
}
\value{
A fitted tidyFit class model.
}
\description{
Fits a Markov-Switching regression and returns the results as a tibble. The function can be used with \code{\link{regress}}.
}
\details{
\strong{Hyperparameters:}

\emph{None. Cross validation not applicable.}

The function provides a wrapper for \code{MSwM::msmFit}. Note that only the regression method with 'lm' is implemented at this stage.

An argument \code{index_col} can be passed, which allows a custom index to be added to \code{coef(m("mslm"))} (e.g. a date index).
}
\examples{
# Load data
data <- tidyfit::Factor_Industry_Returns
data <- dplyr::filter(data, Industry == "HiTec")
data <- dplyr::select(data, -Industry)

ctr <- list(maxiter = 10, parallelization = FALSE)

# Stand-alone function
fit <- m("mslm", Return ~ ., data, index_col = "Date", k = 2, control = ctr)
fit

# Within 'regress' function
fit <- regress(data, Return ~ .,
               m("mslm", index_col = "Date", k = 2, control = ctr))
tidyr::unnest(coef(fit), model_info)

}
\seealso{
\code{\link{.model.tvp}} and \code{\link{m}} methods
}
\author{
Johann Pfitzinger
}
