% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.pcr.R
\name{.model.pcr}
\alias{.model.pcr}
\title{Principal Components Regression for \code{tidyfit}}
\usage{
\method{.model}{pcr}(x = NULL, y = NULL, control = NULL, ...)
}
\arguments{
\item{x}{Input matrix or data.frame, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.}

\item{y}{Response variable.}

\item{control}{Additional arguments passed to \code{pls::pcr}.}

\item{...}{Not used.}
}
\value{
A 'tibble'.
}
\description{
Fits a principal components regression and returns the results as a tibble. The function can be used with \code{\link{regress}}.
}
\details{
\strong{Hyperparameters:}
\itemize{
\item \code{ncomp} \emph{(number of components)}
}

The principal components regression is fitted using \code{pls} package. Covariates are standardized, with coefficients back-transformed to the original scale. An intercept is always included.

If no hyperparameter grid is passed (\code{is.null(control$ncomp)}), the default is \code{seq(1, nvars)}, where nvars is the number of features.

Note that at present \code{pls} does not offer weighted implementations or non-gaussian response. The method can therefore only be used with \code{\link{regress}}
}
\examples{
x <- matrix(rnorm(100 * 20), 100, 20)
y <- rnorm(100)
fit <- m("pcr", x, y, ncomp = 4)
fit

}
\references{
Kristian Hovde Liland, Bjørn-Helge Mevik, Ron Wehrens (2022).
pls: Partial Least Squares and Principal Component Regression.
R package version 2.8-1. URL https://CRAN.R-project.org/package=pls.
}
\seealso{
\code{\link{.model.plsr}} and \code{\link{m}} methods
}
\author{
Johann Pfitzinger
}
