% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.pcr.R
\name{.model.pcr}
\alias{.model.pcr}
\title{Principal Components Regression for \code{tidyfit}}
\usage{
\method{.model}{pcr}(self, data = NULL)
}
\arguments{
\item{self}{a tidyFit R6 class.}

\item{data}{a data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}
}
\value{
A fitted tidyFit class model.
}
\description{
Fits a principal components regression and returns the results as a tibble. The function can be used with \code{\link{regress}}.
}
\details{
\strong{Hyperparameters:}
\itemize{
\item \code{ncomp} \emph{(number of components)}
\item \code{ncomp_pct} \emph{(number of components, percentage of features)}
}

The principal components regression is fitted using \code{pls} package. Covariates are standardized, with coefficients back-transformed to the original scale. An intercept is always included.

If no hyperparameter grid is passed (\code{is.null(control$ncomp) & is.null(control$ncomp_pct)}), the default is \code{ncomp_pct = seq(0, 1, length.out = 20)}, where 0 results in one component and 1 results in the number of features.

Note that at present \code{pls} does not offer weighted implementations or non-gaussian response. The method can therefore only be used with \code{\link{regress}}
}
\examples{
# Load data
data <- tidyfit::Factor_Industry_Returns

# Stand-alone function
fit <- m("pcr", Return ~ ., data, ncomp = 3)
fit

# Within 'regress' function
fit <- regress(data, Return ~ ., m("pcr"), .mask = c("Date", "Industry"), .cv = "vfold")
coef(fit)

}
\references{
Kristian Hovde Liland, Bjørn-Helge Mevik, Ron Wehrens (2022).
pls: Partial Least Squares and Principal Component Regression.
R package version 2.8-1. URL https://CRAN.R-project.org/package=pls.
}
\seealso{
\code{\link{.model.plsr}} and \code{\link{m}} methods
}
\author{
Johann Pfitzinger
}
