% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.quantile.R
\name{.model.quantile}
\alias{.model.quantile}
\title{Quantile regression for \code{tidyfit}}
\usage{
\method{.model}{quantile}(formula = NULL, data = NULL, control = NULL, ...)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted.}

\item{data}{a data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{control}{Additional arguments passed to \code{quantreg::rq}.}

\item{...}{Not used.}
}
\value{
A 'tibble'.
}
\description{
Fits a linear quantile regression and returns the results as a tibble. The function can be used with \code{\link{regress}}.
}
\details{
\strong{Hyperparameters:}

\emph{None. Cross validation not applicable.}

The function provides a wrapper for \code{quantreg::rq}. The argument \code{tau} is the chosen quantile (default \code{tau = 0.5}).
}
\examples{
x <- matrix(rnorm(100 * 20), 100, 20)
y <- rnorm(100)
fit <- m("quantile", x, y, tau = c(0.1, 0.5, 0.9))
fit

}
\seealso{
\code{\link{.model.lm}}, \code{\link{.model.bayes}} and \code{\link{m}} methods
}
\author{
Johann Pfitzinger
}
