% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.robust.R
\name{.model.robust}
\alias{.model.robust}
\title{Robust regression for \code{tidyfit}}
\usage{
\method{.model}{robust}(formula = NULL, data = NULL, control = NULL, ...)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted.}

\item{data}{a data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{control}{Additional arguments passed to \code{MASS::rlm}.}

\item{...}{Not used.}
}
\value{
A 'tibble'.
}
\description{
Fits a robust linear regression and returns the results as a tibble. The function can be used with \code{\link{regress}}.
}
\details{
\strong{Hyperparameters:}

\emph{None. Cross validation not applicable.}

The function provides a wrapper for \code{MASS::rlm}.

An argument \code{vcov.} can be passed in control or to \code{...} in \code{\link{m}} to estimate the model with robust standard errors. \code{vcov.} can be one of "BS", "HAC", "HC" and "OPG" and is passed to the \code{sandwich} package.
}
\examples{
# Load data
data <- tidyfit::Factor_Industry_Returns

# Stand-alone function
fit <- m("robust", Return ~ `Mkt-RF` + HML + SMB, data)
fit

# Within 'regress' function
fit <- regress(data, Return ~ ., m("robust"), .mask = c("Date", "Industry"))
coef(fit)

# With robust standard errors
fit <- m("robust", Return ~ `Mkt-RF` + HML + SMB, data, vcov. = "HAC")
tidyr::unnest(fit, settings)

}
\seealso{
\code{\link{.model.lm}} and \code{\link{m}} methods
}
\author{
Johann Pfitzinger
}
