% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.subset.R
\name{.model.subset}
\alias{.model.subset}
\title{Best subset regression and classification for \code{tidyfit}}
\usage{
\method{.model}{subset}(x = NULL, y = NULL, control = NULL, ...)
}
\arguments{
\item{x}{Input matrix or data.frame, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.}

\item{y}{Response variable.}

\item{control}{Additional arguments passed to \code{bestglm::bestglm}.}

\item{...}{Not used.}
}
\value{
A 'tibble'.
}
\description{
Fits a best subset regression or classification and returns the results as a tibble. The function can be used with \code{\link{regress}} and \code{\link{classify}}.
}
\details{
\strong{Hyperparameters:}

\emph{None. Cross validation not applicable.}

The best subset regression is estimated using \code{bestglm::bestglm} which is a wrapper around \code{leaps::regsubsets} for the regression case, and performs an exhaustive search for the classification case. For classification pass \code{family = "binomial"} to \code{control} or to the argument \code{...} in \code{\link{m}}.

An intercept is always included and features are standardized with coefficients transformed to the original scale.

Forward or backward selection can be performed by passing \code{method = "forward"} or \code{method = "backward"} to \code{\link{m}}.
}
\examples{
x <- matrix(rnorm(100 * 20), 100, 20)
y <- rnorm(100)
fit <- m("subset", x, y)
fit

}
\references{
A.I. McLeod, Changjiang Xu and Yuanhao Lai (2020).
bestglm: Best Subset GLM and Regression Utilities.
R package version 0.37.3. URL https://CRAN.R-project.org/package=bestglm.
}
\seealso{
\code{\link{m}} method
}
\author{
Johann Pfitzinger
}
