% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\name{m}
\alias{m}
\title{Generic model wrapper for \code{tidyfit}}
\usage{
m(model_method, formula = NULL, data = NULL, ...)
}
\arguments{
\item{model_method}{The name of the method to fit. See Details.}

\item{formula}{an object of class "formula": a symbolic description of the model to be fitted.}

\item{data}{a data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{...}{Additional arguments passed to the underlying method function (e.g. \code{lm} or \code{glm}).}
}
\value{
A 'tidyfit.models' data frame.
}
\description{
The function can fit various regression or classification models and returns the results as a tibble. \code{m()} can be used in conjunction with \code{\link{regress}} and \code{\link{classify}}, or as a stand-alone function.
}
\details{
\code{model_method} specifies the model to fit to the data and can take one of several options:
\subsection{Linear (generalized) regression or classification}{

\code{"lm"} performs an OLS regression using \code{stats::lm}. See \code{\link{.model.lm}} for details.

\code{"glm"} performs a generalized regression or classification using \code{stats::glm}. See \code{\link{.model.glm}} for details.

\code{"robust"} performs a robust regression using \code{MASS::rlm}. See \code{\link{.model.robust}} for details.

\code{"quantile"} performs a quantile regression using \code{quantreg::rq}. See \code{\link{.model.quantile}} for details.
}

\subsection{Regression and classification with L1 and L2 penalties}{

\code{"lasso"} performs a linear regression or classification with L1 penalty using \code{glmnet::glmnet}. See \code{\link{.model.lasso}} for details.

\code{"ridge"} performs a linear regression or classification with L2 penalty using \code{glmnet::glmnet}. See \code{\link{.model.lasso}} for details.

\code{"adalasso"} performs an Adaptive Lasso regression or classification using \code{glmnet::glmnet}. See \code{\link{.model.adalasso}} for details.

\code{"enet"} performs a linear regression or classification with L1 and L2 penalties using \code{glmnet::glmnet}. See \code{\link{.model.enet}} for details.
}

\subsection{Gradient boosting}{

\code{"boost"} performs gradient boosting regression or classification using \code{mboost::glmboost}. See \code{\link{.model.boost}} for details.
}

\subsection{Factor regressions}{

\code{"pcr"} performs a principal components regression using \code{pls::pcr}. See \code{\link{.model.pcr}} for details.

\code{"plsr"} performs a partial least squares regression using \code{pls::plsr}. See \code{\link{.model.plsr}} for details.

\code{"hfr"} performs a hierarchical feature regression using \code{hfr::hfr}. See \code{\link{.model.hfr}} for details.
}

\subsection{Best subset selection}{

\code{"subset"} performs a best subset regression or classification using \code{bestglm::bestglm} (wrapper for \code{leaps}). See \code{\link{.model.subset}} for details.
}

\subsection{Bayesian methods}{

\code{"bayes"} performs a Bayesian generalized regression or classification using \code{arm::bayesglm}. See \code{\link{.model.bayes}} for details.

\code{"tvp"} performs a Bayesian time-varying parameter regression using \code{shrinkTVP::shrinkTVP}. See \code{\link{.model.tvp}} for details.
}

\subsection{Mixed-effects modeling}{

\code{"glmm"} performs a mixed-effects GLM using \code{lme4::glmer}. See \code{\link{.model.glmm}} for details.
}

\subsection{Specialized time series methods}{

\code{"mslm"} performs a Markov-switching regression using \code{MSwM::msmFit}. See \code{\link{.model.mslm}} for details.

When called without \code{formula} and \code{data} arguments, the function returns a 'tidyfit.models' data frame with unfitted models.
}
}
\examples{
# Load data
data <- tidyfit::Factor_Industry_Returns

# Stand-alone function
fit <- m("lm", Return ~ ., data)
fit

# Within 'regress' function
fit <- regress(data, Return ~ ., m("lm"), .mask = "Date")
fit

}
\seealso{
\code{\link{regress}} and \code{\link{classify}} methods
}
\author{
Johann Pfitzinger
}
