% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_imp.tidyfit.models.R
\name{var_imp.tidyfit.models}
\alias{var_imp.tidyfit.models}
\title{Variable importance measures for a \code{tidyfit.models} frame}
\usage{
\method{var_imp}{tidyfit.models}(object, method = NULL, ..., .keep_grid_id = FALSE)
}
\arguments{
\item{object}{\code{model.frame} created using \code{\link{regress}}, \code{\link{classify}} or \code{\link{m}}}

\item{method}{the variable importance method used to create explanations. See 'Details' for possible options.}

\item{...}{additional arguments passed to the importance method}

\item{.keep_grid_id}{boolean. By default the grid ID column is dropped, if there is only one unique setting per model or group. \code{.keep_grid_id = TRUE} ensures that the column is never dropped.}
}
\value{
A 'tibble'.
}
\description{
The function generates explanations for all models in a \code{tidyfit.models} frame and outputs a tidy frame.
}
\details{
The function uses the 'model_object' column in a \code{tidyfit.model} frame to return variable importance measures for each model.

\strong{Possible methods include:}
\subsection{Linear regression ('lm'):}{
\itemize{
\item 'shapley_reg' for Shapley regression (default). The method uses the 'lmg' algorithm implemented in the \code{relaimpo}-package. Other algorithms can be applied by explicitly passing 'type' to \code{...}, which is passed to \code{relaimpo::calc.relimp}.
\item 'rel_weights' for relative weights. The method uses the 'genizi' algorithm implemented in the \code{relaimpo}-package.
}
}
}
\examples{
data <- dplyr::group_by(tidyfit::Factor_Industry_Returns, Industry)
fit <- regress(data, Return ~ ., m("lm"), .mask = "Date")
var_imp(fit, method = "rel_weights")

}
\seealso{
\code{\link{coef.tidyfit.models}}, \code{\link{residuals.tidyfit.models}} and \code{\link{fitted.tidyfit.models}}
}
\author{
Johann Pfitzinger
}
